vinyl-catalog/sidecar/requirements.txt
fastapi
uvicorn[standard]
easyocr
pillow
numpy
openai
google-generativeai
python-dotenv
requests
